requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0

# Database connectors
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0

# Data analysis and visualization
jupyter>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional: Data validation
great-expectations>=0.15.0

# Optional: SQL utilities
sqlalchemy>=1.4.0
